apps/food-diary-api/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
josepy==2.0.0
mozilla-django-oidc==4.0.1
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
